+++ HydroGarden.Foundation.Abstractions/HydroGarden.Foundation.Abstractions.csproj
@@ -6,8 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Interfaces\Events\" />
-  </ItemGroup>
-
 </Project>
