+++ HydroGarden.Foundation.Abstractions/Interfaces/Events/IComponentConnection.cs
