+++ HydroGarden.Foundation.Abstractions/Interfaces/Events/IEventRoutingData.cs
@@ -1,5 +1,4 @@
-﻿
-namespace HydroGarden.Foundation.Abstractions.Interfaces
+﻿namespace HydroGarden.Foundation.Abstractions.Interfaces.Events
 {
     /// <summary>
     /// Interface for event routing data
