+++ HydroGarden.Foundation.Abstractions/Interfaces/Events/IEventSubscription.cs
@@ -1,4 +1,4 @@
-﻿namespace HydroGarden.Foundation.Abstractions.Interfaces
+﻿namespace HydroGarden.Foundation.Abstractions.Interfaces.Events
 {
     public interface IEventSubscription
     {
