+++ HydroGarden.Foundation.Abstractions/Interfaces/Events/IHydroGardenEvent.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HydroGarden.Foundation.Abstractions.Interfaces
+namespace HydroGarden.Foundation.Abstractions.Interfaces.Events
 {
     /// <summary>
     /// Base interface for all HydroGarden events
