+++ HydroGarden.Foundation.Common/Events/ConcreteEvents.cs
@@ -1,4 +1,6 @@
 ﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+using HydroGarden.Foundation.Abstractions.Interfaces.Components;
+using HydroGarden.Foundation.Abstractions.Interfaces.Events;
 
 
 namespace HydroGarden.Foundation.Common.Events
