+++ HydroGarden.Foundation.Common/Events/EventRoutingData.cs
@@ -1,4 +1,4 @@
-﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+﻿using HydroGarden.Foundation.Abstractions.Interfaces.Events;
 
 namespace HydroGarden.Foundation.Common.Events
 {
