+++ HydroGarden.Foundation.Common/Events/EventSubscription.cs
@@ -1,4 +1,5 @@
 ﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+using HydroGarden.Foundation.Abstractions.Interfaces.Events;
 
 
 namespace HydroGarden.Foundation.Common.Events
