+++ HydroGarden.Foundation.Common/Logging/HydroGardenLogger.cs
@@ -1,4 +1,4 @@
-﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+﻿using HydroGarden.Foundation.Abstractions.Interfaces.Logging;
 using System;
 using System.IO;
