+++ HydroGarden.Foundation.Core/Components/Devices/PumpDevice.cs
@@ -1,4 +1,4 @@
-﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+﻿using HydroGarden.Foundation.Abstractions.Interfaces.Logging;
 using HydroGarden.Foundation.Common.Logging;
 using System;
 using System.Threading;
