+++ HydroGarden.Foundation.Core/Services/PersistentService.cs
@@ -1,4 +1,7 @@
 ﻿using HydroGarden.Foundation.Abstractions.Interfaces;
+using HydroGarden.Foundation.Abstractions.Interfaces.Components;
+using HydroGarden.Foundation.Abstractions.Interfaces.Logging;
+using HydroGarden.Foundation.Abstractions.Interfaces.Services;
 using HydroGarden.Foundation.Common.Caching;
 using HydroGarden.Foundation.Common.Logging;
 using HydroGarden.Foundation.Core.Components;
